+++ package.json
@@ -74,7 +74,7 @@
     "nyc": "^15.1.0",
     "prettier": "^2.2.1",
     "source-map-support": "^0.5.19",
-    "ts-jest": "^26.4.4",
+    "ts-jest": "^26.5.6",
     "ts-loader": "^8.0.14",
     "ts-node": "^8.10.2",
     "typescript": "^4.1.3",
