+++ src/utils/addon-api.ts
@@ -14,16 +14,8 @@ import { log, logInfo, logError, safeStringify } from './logger';
 import Server from '../server';
 import ASTPath from './../glimmer-utils';
 import DAGMap from 'dag-map';
-import CoreScriptDefinitionProvider from './../builtin-addons/core/script-definition-provider';
-import CoreTemplateDefinitionProvider from './../builtin-addons/core/template-definition-provider';
-import ScriptCompletionProvider from './../builtin-addons/core/script-completion-provider';
-import TemplateCompletionProvider from './../builtin-addons/core/template-completion-provider';
+
 import { Project } from '../project';
-
-import TemplateLintFixesCodeAction from '../builtin-addons/core/code-actions/template-lint-fixes';
-import TemplateLintCommentsCodeAction from '../builtin-addons/core/code-actions/template-lint-comments';
-import TypedTemplatesCodeAction from '../builtin-addons/core/code-actions/typed-template-comments';
-
 interface BaseAPIParams {
   server: Server;
   textDocument: TextDocumentIdentifier;
@@ -96,38 +88,6 @@ export async function queryELSAddonsAPIChain(callbacks: any[], root: string, par
   }
 
   return lastResult;
-}
-
-export function initBuiltinProviders(): ProjectProviders {
-  const scriptDefinition = new CoreScriptDefinitionProvider();
-  const templateDefinition = new CoreTemplateDefinitionProvider();
-  const scriptCompletion = new ScriptCompletionProvider();
-  const templateCompletion = new TemplateCompletionProvider();
-
-  const templateLintFixesCodeAction = new TemplateLintFixesCodeAction();
-  const templateLintCommentsCodeAction = new TemplateLintCommentsCodeAction();
-  const typedTemplatesCodeAction = new TypedTemplatesCodeAction();
-
-  return {
-    definitionProviders: [scriptDefinition.onDefinition.bind(scriptDefinition), templateDefinition.onDefinition.bind(templateDefinition)],
-    referencesProviders: [],
-    codeActionProviders: [
-      templateLintFixesCodeAction.onCodeAction.bind(templateLintFixesCodeAction),
-      templateLintCommentsCodeAction.onCodeAction.bind(templateLintCommentsCodeAction),
-      typedTemplatesCodeAction.onCodeAction.bind(typedTemplatesCodeAction),
-    ],
-    initFunctions: [
-      templateLintFixesCodeAction.onInit.bind(templateLintFixesCodeAction),
-      templateLintCommentsCodeAction.onInit.bind(templateLintCommentsCodeAction),
-      typedTemplatesCodeAction.onInit.bind(typedTemplatesCodeAction),
-      templateCompletion.initRegistry.bind(templateCompletion),
-      scriptCompletion.initRegistry.bind(scriptCompletion),
-      templateDefinition.onInit.bind(templateDefinition),
-    ],
-    info: [],
-    addonsMeta: [],
-    completionProviders: [scriptCompletion.onComplete.bind(scriptCompletion), templateCompletion.onComplete.bind(templateCompletion)],
-  };
 }
 
 export function isConstructor(obj: any) {
